move session init into bootstrap
