+++ cmd/slackdump/internal/archive/search.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	"github.com/rusq/fsadapter"
+	"github.com/rusq/slackdump/v3/cmd/slackdump/internal/bootstrap"
 	"github.com/rusq/slackdump/v3/cmd/slackdump/internal/cfg"
 	"github.com/rusq/slackdump/v3/cmd/slackdump/internal/golang/base"
 	"github.com/rusq/slackdump/v3/internal/chunk"
@@ -109,7 +110,7 @@ func initController(ctx context.Context, args []string) (*control.Controller, fu
 		return nil, nil, errNoOutput
 	}
 
-	sess, err := cfg.SlackdumpSession(ctx)
+	sess, err := bootstrap.SlackdumpSession(ctx)
 	if err != nil {
 		base.SetExitStatus(base.SInitializationError)
 		return nil, nil, err
