+++ cmd/slackdump/internal/bootstrap/slackdump.go
@@ -1,10 +1,12 @@
-package cfg
+package bootstrap
 
 import (
 	"context"
 
 	"github.com/rusq/slackdump/v3"
 	"github.com/rusq/slackdump/v3/auth"
+	"github.com/rusq/slackdump/v3/cmd/slackdump/internal/cfg"
+	"github.com/rusq/slackdump/v3/cmd/slackdump/internal/workspace"
 )
 
 // SlackdumpSession returns the Slackdump Session initialised with the provider
@@ -18,9 +20,9 @@ func SlackdumpSession(ctx context.Context, opts ...slackdump.Option) (*slackdump
 	}
 
 	var stdOpts = []slackdump.Option{
-		slackdump.WithLogger(Log),
-		slackdump.WithForceEnterprise(ForceEnterprise),
-		slackdump.WithLimits(Limits),
+		slackdump.WithLogger(cfg.Log),
+		slackdump.WithForceEnterprise(cfg.ForceEnterprise),
+		slackdump.WithLimits(cfg.Limits),
 	}
 
 	stdOpts = append(stdOpts, opts...)
@@ -29,4 +31,13 @@ func SlackdumpSession(ctx context.Context, opts ...slackdump.Option) (*slackdump
 		prov,
 		stdOpts...,
 	)
+}
+
+// CurrentProviderCtx returns the context with the current provider.
+func CurrentProviderCtx(ctx context.Context) (context.Context, error) {
+	prov, err := workspace.AuthCurrent(ctx, cfg.CacheDir(), cfg.Workspace, cfg.LegacyBrowser)
+	if err != nil {
+		return ctx, err
+	}
+	return auth.WithContext(ctx, prov), nil
 }
