+++ cmd/slackdump/internal/bootstrap/slackdump_test.go
@@ -1,4 +1,4 @@
-package cfg
+package bootstrap
 
 import (
 	"context"
